requirements.txt
django
djangorestframework
django-cors-headers
markdown
pytz
gunicorn==19.7.1
django-filter
flup
django.conf.urls.defaults
mysql-connector-python
